refactor: make custom hook use kebab case
